% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{add_tweet_kind_cols}
\alias{add_tweet_kind_cols}
\title{Augment \code{rtweet} data.frame}
\usage{
add_tweet_kind_cols(data = NULL)
}
\arguments{
\item{data}{data.frame (created using \code{rtweet} package).}
}
\value{
data.frame
}
\description{
Add columns to \code{rtweet} data.frame for 'kind' data.
}
\details{
The columns added include
'hashtag', 'link', 'rt', 'quote', and 'reply', all of which are logicals.
A 'type' column is also added--it is a factored version of 'rt', 'quote', and 'reply'.
}
\seealso{
\url{https://juliasilge.com/blog/ten-thousand-data/}.
}
