% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{append_dummy_cols}
\alias{append_dummy_cols}
\title{Post-process data in \code{wrapper_func()}}
\usage{
append_dummy_cols(data = NULL, num_cols_expect = NULL, add = TRUE)
}
\arguments{
\item{data}{data.frame.}

\item{num_cols_expect}{numeric.}

\item{add}{logical. whether or not to try to add a dummy column. Only
applied if the difference between the expected and actual number of columns is 1
(where the expected is greater than the actual).}
}
\value{
data.frame.
}
\description{
Add dummy columns to \code{data}, if necessary.
}
\details{
To be used exclusively by \code{wrapper_func()}.
}
