% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{clean_tweets_at}
\alias{clean_tweets_at}
\alias{clean_tweets}
\title{Clean (and augment) \code{rtweet} data.frame}
\usage{
clean_tweets_at(data = NULL, facet, trim = TRUE, cols = c("status_id",
  "created_at", "user_id", "screen_name", "text", "display_text_width",
  "reply_to_status_id", "is_quote", "is_retweet", "favorite_count",
  "retweet_count", "hashtags", "symbols", "urls_url", "urls_expanded_url",
  "media_expanded_url", "ext_media_expanded_url"),
  timezone = "America/Chicago")

clean_tweets(data = NULL, facet, trim = TRUE, cols = c("status_id",
  "created_at", "user_id", "screen_name", "text", "display_text_width",
  "reply_to_status_id", "is_quote", "is_retweet", "favorite_count",
  "retweet_count", "hashtags", "symbols", "urls_url", "urls_expanded_url",
  "media_expanded_url", "ext_media_expanded_url"),
  timezone = "America/Chicago")
}
\arguments{
\item{data}{data.frame (created using \code{rtweet} package).}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} used for facetting.
Does not need to be specified.
Included in \code{cols} if specified.}

\item{trim}{logical. Indicates whether or not to select only certain columns
(and drop the others).}

\item{cols}{character (vector). Name(s) of column(s) in \code{data} to keep.
Only relevant if \code{trim = TRUE}.}

\item{timezone}{character. Passed directly to \code{lubridate::with_tz()} as \code{tzone} parameter.}
}
\value{
data.frame
}
\description{
Munge an \code{rtweet} data.frame (according to personal preferences)
for subsequent analysis.
}
\details{
Converts nested lists to character(s).
Adds a \code{timestamp} column that is derived from the \code{created_at}
column. Also, adds a \code{time} column that represents the hour in the day of \code{created_at}.
}
\seealso{
\url{https://juliasilge.com/blog/ten-thousand-data/}.
\url{https://buzzfeednews.github.io/2018-01-trump-twitter-wars/}.
}
