% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{compute_change_at}
\alias{compute_change_at}
\title{Compute change}
\usage{
compute_change_at(data = NULL, timebin = NULL, word = "word",
  timefloor = NULL, top_pct = 0.05, return_models = TRUE,
  return_data = FALSE, return_both = FALSE)
}
\arguments{
\item{data}{data.frame.}

\item{timebin}{character. Name of column in \code{data} specifying temporal period
to use to compute change.}

\item{word}{character. Name of column in \code{data} corresponding to n-gram.}

\item{timefloor}{character. Passed directly to \code{lubridate::floor_date()} as\code{unit} parameter.}

\item{top_pct}{numeric. Number between 0 and 1. Default is provided.}

\item{return_models}{logical. Whether to return just the models. This is probably the preferred
option when calling \code{compute_change_at()} directly. Default is provided.}

\item{return_data}{logical. Whether to 'bytime' data which is used as
the \code{data} parameter in \code{stats::glm()} for creating models. Default is provided.
Needed when using \code{visualize_change_at()}.}

\item{return_both}{logical. Set to \code{TRUE} when \code{visualize_change_at()}. Default is provided.}
}
\value{
data.frame.
}
\description{
Compute change in word usage over time.
}
\details{
None.
}
\seealso{
\url{https://www.tidytextmining.com/twitter.html#changes-in-word-use}
}
