% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratio.R
\name{compute_logratios_facet_by2_at}
\alias{compute_logratios_facet_by2_at}
\alias{compute_logratios_facet_by2}
\title{Compute token log ratios in pairs}
\usage{
compute_logratios_facet_by2_at(data = NULL,
  func = compute_logratios_facet_wide_at, xy_grid = NULL, xy_nms = NULL,
  facet = NULL, ...)

compute_logratios_facet_by2(data = NULL,
  func = compute_logratios_facet_wide_at, xy_grid = NULL, xy_nms = NULL,
  facet = NULL, ...)
}
\arguments{
\item{data}{data.frame. Data to apply a function over.}

\item{func}{function. Wrapped by \code{do.call()}.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} corresponding to group
by which count is made.}

\item{...}{dots. Adddition parameters to pass to \code{compute_logratios_facet_wide_at()}.}
}
\value{
data.frame.
}
\description{
Apply \code{compute_logratios_facet_wide_at()} to
\code{data} in pairs of \code{facet} values.
}
\details{
Calls \code{wrapper_func} internally.
}
