% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratio.R
\name{compute_logratios_multi_wide_at}
\alias{compute_logratios_multi_wide_at}
\alias{compute_logratios_multi_wide}
\title{Compute log ratios}
\usage{
compute_logratios_multi_wide_at(data = NULL, word = "word", multi = NULL,
  cnt_min = 10)

compute_logratios_multi_wide(data = NULL, word = "word", multi = NULL,
  cnt_min = 10)
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for count. Default is provided}

\item{multi}{character. Name of column in \code{data} corresponding to group
by which count is made.}

\item{cnt_min}{numeric. If not null, serves as a filter for which to compute log ratios.}
}
\value{
data.frame.
}
\description{
Compute log ratios for usage of n-grams between two entities.
}
\details{
Note that an attempt was made to make this function flexible enough,
to consider a \code{num_top} or \code{top_pct} parameter that might be used if
\code{cnt_min = NULL}, but this can be troublesome because \code{do.call()}
does not seem to handle \code{NULL} in \code{...}.
}
\seealso{
\url{https://www.tidytextmining.com/twitter.html#comparing-word-usage}.
\url{http://varianceexplained.org/r/trump-tweets/}.
}
