% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sents.R
\name{compute_sent_summary_multi_at}
\alias{compute_sent_summary_multi_at}
\alias{compute_sent_summary_multi}
\alias{compute_sent_summary_at}
\alias{compute_sent_summary}
\title{Compute sentiment scores.}
\usage{
compute_sent_summary_multi_at(data = NULL, word = "word", multi = NULL,
  feature = NULL, lexicon = "bing", ...)

compute_sent_summary_multi(data = NULL, word = "word", multi = NULL,
  feature = NULL, lexicon = "bing", ...)

compute_sent_summary_at(data = NULL, word, multi, feature, ...)

compute_sent_summary(data = NULL, word, multi, feature, ...)
}
\arguments{
\item{data}{data.frame. Must already be in 'tidy' format.}

\item{word}{character. Name of column in \code{data} whose values are scored.}

\item{multi}{character. Name of column in \code{data} used for grouping in the case
that there are more than one entitiy.}

\item{feature}{character. Name of column in \code{data} whose values serve as
entity across which 'word' values are scored. Probably something like 'sentence',
or, in the case of Twitter data, 'status_id'.}

\item{lexicon}{character. Name of lexicon available in \code{tidytext} package to use.}

\item{...}{dots. Parameters to pass to \code{get_sents()} internally.}
}
\value{
data.frame.
}
\description{
Compute sentiment scores given n-grams.
}
\details{
Heavily influenced by \href{https://www.tidytextmining.com/}{\emph{Text Mining with R}}.
Creates output columns 'sentiment' and 'score'.
}
\seealso{
Don't have an explicit url.
}
