% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sents.R
\name{compute_sentratios_facet_by2_at}
\alias{compute_sentratios_facet_by2_at}
\alias{compute_sentratios_facet_by2}
\title{Compute token sentiment log ratios in pairs}
\usage{
compute_sentratios_facet_by2_at(data = NULL,
  func = compute_logratios_facet_wide_at, token = NULL, facet = NULL,
  xy_grid, xy_nms, cnt_min = 10, lexicon = "bing", ...)

compute_sentratios_facet_by2(data = NULL,
  func = compute_logratios_facet_wide_at, token = NULL, facet = NULL,
  xy_grid, xy_nms, cnt_min = 10, lexicon = "bing", ...)
}
\arguments{
\item{data}{data.frame. Data to apply a function over.}

\item{func}{function. Wrapped by \code{do.call()}.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use for count. Default is provided}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} corresponding to group
by which count is made.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{cnt_min}{numeric. If not null, serves as a filter for which to compute log ratios.}

\item{lexicon}{character. Specifies lexicon to use. Must be a lexicon available in the \code{tidytext::package}.}

\item{...}{dots. Adddition parameters to pass to \code{get_sents()}.}
}
\value{
data.frame.
}
\description{
Apply \code{compute_logratios_facet_wide_at()} to
\code{data} in pairs of \code{facet} values, then do post-processing for sentiments.
}
\details{
Calls \code{wrapper_func} internally. This function is very
similar to \code{compute_sentratios_facet_by2_at()}, but with some post-processing.
Note that \code{...} are passed to \code{get_sents()} here (instead of to
\code{compute_logratios_facet_wide_at()}, so \code{token} and \code{cnt_min} must
be checked for NULL explicilty.
}
