% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{compute_tfidf_at}
\alias{compute_tfidf_at}
\alias{compute_tfidf}
\title{Compute TF-IDF}
\usage{
compute_tfidf_at(data = NULL, token = NULL, doc = NULL)

compute_tfidf(..., token, doc)
}
\arguments{
\item{data}{data.frame.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use
as \code{term} in \code{tidytext::bind_tf_idf()}.}

\item{doc}{bare for NSE; character for SE. Name of column in \code{data} to use
as \code{document} in \code{tidytext::bind_tf_idf()}.}

\item{...}{dots. Parameters to pass directly to \code{visualize_time()}.}
}
\value{
data.frame.
}
\description{
Compute term-frequency, inverse-document-frequency.
}
\details{
Calls \code{tidytext::bind_tf_idf()} internally. Note that the
output 'tfidf' column is named \code{tf_idf} because this is the convention
of the \code{tidytext} package.
}
