% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{compute_time_elapsed}
\alias{compute_time_elapsed}
\title{Compute time difference}
\usage{
compute_time_elapsed(date_start = NULL, date_end = NULL, type = c("year",
  "month", "day", "hour"))
}
\arguments{
\item{date_start}{Date-time (or character that can be converted to Date-time).}

\item{date_end}{Date-time (or character that can be converted to Date-time).}

\item{type}{character. 'unit' to use to calculate time difference.
One of 'year', 'month', 'day', or 'hour'.}
}
\value{
numeric.
}
\description{
Computes the difference in time between two Date-time values
in a specified time 'unit'.
}
\details{
Called in \code{compute_timefilter_multi_at()}.
}
\seealso{
https://stackoverflow.com/
}
