% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{compute_timefilter_multi_at}
\alias{compute_timefilter_multi_at}
\title{Add time-related columns}
\usage{
compute_timefilter_multi_at(data = NULL, timebin = NULL, multi = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{timebin}{character. Name of column in \code{data} to use for time axis.}

\item{multi}{character. Name of column in \code{data} used for facetting.}
}
\value{
list
}
\description{
Adds useful columns for subsequent time calculations and filtering.
}
\details{
Adds columns for years, months, days, and hours elapsed.
(Calls in \code{compute_time_elapsed()} multiple times.)
Returns list, where \code{data} is augmented data.frame,
\code{date_start} is the very LAST Date-time value, and \code{date_end}
is the very FIRST Date-time value. The \code{date_start} and \code{date_end}
values are defined in this manner such that all data for each 'multi' variable
fits in a singular frame (i.e. there are not 'gaps' where data exists
for one 'multi' variable and not all others). This determination assumes
that the data is continuous.
}
