% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{create_xy_grid}
\alias{create_xy_grid}
\title{Create a 'xy grid'}
\usage{
create_xy_grid(xy_nms)
}
\arguments{
\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}
}
\value{
data.frame.
}
\description{
Create a grid of each comibination of the values in the input vector.
}
\details{
This function works similarly to \code{expand.grid()}.
}
