% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default_facet}
\alias{default_facet}
\title{Default facet parameters}
\usage{
default_facet(facet, scales = "free", ncol = 3, nrow = NULL,
  strip.position = "right", ...)
}
\arguments{
\item{facet, scales, ncol, nrow, strip.position, ...}{Parameters passed directly to
\code{ggplot2::facet_wrap()}}
}
\value{
list
}
\description{
None.
}
\details{
Not exactly analogous to \code{default_theme()} and \code{default_labs()}
because this function returns a list, not a function.
}
