% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default_filter_facet}
\alias{default_filter_facet}
\title{Default word cloud parameters}
\usage{
default_filter_facet(facet_main = NULL, x_include = NULL,
  y_include = NULL, x_exclude = NULL, y_exclude = NULL)
}
\arguments{
\item{facet_main}{of same type as \code{facet} values. Value of single 'main' facet.}

\item{x_include, y_include, x_exclude, y_exclude}{character (vector).
\code{facet} values to include and exclude. Not used if \code{filter_facet = FALSE}.}
}
\value{
list.
}
\description{
None.
}
\details{
Similar to other 'default' functions.
}
