% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default_labs}
\alias{default_labs}
\title{Default label parameters}
\usage{
default_labs(title = NULL, subtitle = NULL, caption = NULL, x = NULL,
  y = NULL)
}
\arguments{
\item{title, subtitle, caption, x, y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}
}
\value{
function.
}
\description{
None.
}
\details{
Plays a similar role as \code{default_theme()}, but for labels.
}
