% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default_scale_manual}
\alias{default_scale_manual}
\title{Default word cloud parameters}
\usage{
default_scale_manual(values = "grey50", ...)
}
\arguments{
\item{values}{character (vector). Passed directly to \code{ggplot2::scale_} function.}

\item{...}{dots. Additional parameters to pass to \code{ggplot2::scale_} function.}
}
\value{
list.
}
\description{
None.
}
\details{
Similar to other 'default' functions.
}
