% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default_wordcloud}
\alias{default_wordcloud}
\title{Default word cloud parameters}
\usage{
default_wordcloud(data, token, colors = "grey50", max.words = 50,
  random.order = FALSE, ...)
}
\arguments{
\item{data}{data.frame. Used for computing counts for \code{token} to use for
\code{freq} argument in \code{wordcloud::wordcloud()}.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} used for
\code{words} argument in \code{wordcloud::wordcloud()}.}

\item{colors, max.words, random.order, ...}{Parameters passed directly to
\code{ggplot2::facet_wrap()}}
}
\value{
list
}
\description{
None.
}
\details{
Not exactly analogous to \code{default_theme()} and \code{default_labs()}
because this function returns a list, not a function. More similar to
\code{default_facet()}.
}
