% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ngrams_freqs_byx.R
\name{get_ngrams_freqs_byx}
\alias{get_ngrams_freqs_byx}
\title{Compute n-gram frequency}
\usage{
get_ngrams_freqs_byx(data = NULL, colname_x = NULL, colname_cnt = "word")
}
\arguments{
\item{data}{data.frame.}

\item{colname_x}{character. Input column name of 'facetting' variable. Probably something like 'name'.}

\item{colname_cnt}{character. Input column name of value to count. Default: 'word'}
}
\value{
data.frame.
}
\description{
Compute the frequency of n-grams.
}
\details{
Heavily influence by the _Tidy Text Mining with R_ book.
}
