% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sents.R
\name{get_sents}
\alias{get_sents}
\title{Get sentiment dictionary}
\usage{
get_sents(lexicon = NULL, normalize = TRUE)
}
\arguments{
\item{lexicon}{character. Specifies lexicon to use. Must be a lexicon available in the \code{tidytext::package}.}

\item{normalize}{logical. Only releveant if \code{lexicon = 'afinn'}. Renormalizes
'afinn' scores to -1 to 1 range.}
}
\value{
data.frame
}
\description{
Get words associated with a lexicon for subsequent sentiment analysis.
}
\details{
Called by \code{compute_sent_multi_summary()}.
This function is essentially a customized version of  \code{tidytext::get_sentiments()}.
}
