% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{get_xy_info}
\alias{get_xy_info}
\title{Get information from a particular row in the 'xy grid'}
\usage{
get_xy_info(xy_grid = NULL, xy_nms = NULL, i = NULL)
}
\arguments{
\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{xy_nms}{character (vector). Value in \code{xy_grid} to filter for.
This is not really necessary if \code{xy_grid} has a column corresponding to \code{i}.}

\item{i}{numeric. Used to extract a specific value in \code{xy_nms}.}
}
\value{
list.
}
\description{
Filter the xy grid of each comibination of the values in the input vector.
}
\details{
To be used exclusively by \code{wrapper_func()}.
}
