% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{labs_tetext}
\alias{labs_tetext}
\alias{facet_tetext}
\alias{wordcloud_tetext}
\title{Default label parameters}
\usage{
labs_tetext(title = NULL, subtitle = NULL, caption = NULL, x = NULL,
  y = NULL)

facet_tetext(facet, scales = "free", ncol = 3, nrow = NULL,
  strip.position = "right", ...)

wordcloud_tetext(data, token, colors = "grey50", max.words = 50,
  random.order = FALSE, ...)
}
\arguments{
\item{title, subtitle, caption, lab_x, lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{facet, scales, ncol, nrow, strip.position, ...}{Parameters passed directly to
\code{ggplot2::facet_wrap()}}

\item{data}{data.frame. Used for computing counts for \code{token} to use for
\code{freq} argument in \code{wordcloud::wordcloud()}.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} used for
\code{words} argument in \code{wordcloud::wordcloud()}.}

\item{colors, max.words, random.order, ...}{Parameters passed directly to
\code{ggplot2::facet_wrap()}}
}
\value{
function.

list

list
}
\description{
None.

None.

None.
}
\details{
Plays a similar role as \code{theme_tetext()}, but for labels.

Not exactly analogous to \code{theme_tetext()} and \code{labs_tetext()}
because this function returns a list, not a function.

Not exactly analogous to \code{theme_tetext()} and \code{labs_tetext()}
because this function returns a list, not a function. More similar to
\code{facet_tetext()}.
}
