% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim_bytime_at}
\alias{trim_bytime_at}
\alias{trim_bytime}
\title{Trim data.frame by time}
\usage{
trim_bytime_at(data = NULL, timebin = NULL, start, end, ...)

trim_bytime(data = NULL, timebin = NULL, start, end, ...)
}
\arguments{
\item{data}{data.frame.}

\item{timebin}{character. Name of columin in \code{data} to use for time filtering.}

\item{start, end}{Date-time. If either is missing, then code{compute_timefilter_at()} is called.}

\item{...}{dots. Additional parameters to pass to \code{compute_timefilter_at()}
in the case that start and end are not specified.}
}
\value{
data.frame.
}
\description{
Trim a data.frame to 'fit' within specified 'start' and 'end' Date-time values.
}
\details{
Should be used to a trim data.frame appropriately/dynamically
given an unknown data set where visualization across a single, appropriate time period
is desired. (The unkown data set may have different 'max' and 'min' times
for each 'multi' column.
}
