% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim_data_bytime_at}
\alias{trim_data_bytime_at}
\title{Trim data.frame by time}
\usage{
trim_data_bytime_at(data = NULL, timebin = NULL, start = NULL,
  end = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{timebin}{character. Name of columin in \code{data} to use for time filtering.}

\item{start}{Date-time.}

\item{end}{Date-time.}
}
\value{
data.frame.
}
\description{
Trim a data.frame to 'fit' within specified 'start' and 'end' Date-time values.
}
\details{
Should be used to a trim data.frame appropriately/dynamically
given an unknown data set where visualization across a single, appropriate time period
is desired. (The unkown data set may have different 'max' and 'min' times
for each 'multi' column.
}
