% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs-bigrams.R
\name{visualize_bigram_freqs_at}
\alias{visualize_bigram_freqs_at}
\alias{visualize_bigram_freqs}
\title{Visualize bigrams}
\usage{
visualize_bigram_freqs_at(..., token = NULL, facet = NULL, color = facet,
  num_top = 3, scale_manual_base = default_scale_manual(),
  scale_manual_params = list(), labs_base = default_labs(),
  labs_params = list(title = "Most Frequently Used Word Pairs"),
  theme_base = default_theme(), theme_params = list())

visualize_bigram_freqs(..., token, color, facet)
}
\arguments{
\item{...}{dots. Parameters to pass directly to \code{visualize_time()}.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use as unit for counting.
Probably something like 'word' or 'bigram'.}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} to use for facetting.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{num_top}{numeric. Number of words to show. If between 0 and 1, then assumed to be a percentage.}

\item{scale_manual_base}{list. Parameters to pass to \code{ggplot2::scale_fill_manual()} or
\code{ggplot2::scale_color_manual()}, depending on the aesthetic mapping. A default \code{values}
argument is specified in the internal
function \code{default_scale_manual()},
so if overwriting explicitly, then \code{values} should be included in the list.}

\item{scale_manual_params}{list. Additional parameters to pass to \code{ggplot2::scale_fill_manual()}
or \code{ggplot2::scale_color_manual()} (e.g. \code{breaks}, \code{labels}, or \code{name}).
The default \code{values} specified by \code{scale_manual_base = default_scale_manual()}
can be overwritten a \code{values} argument in this list (or directly with a
\code{values} argument in the \code{scale_manual_base} list.}

\item{labs_base}{\code{ggplot2::labs()} function. Defaults to a pre-determined set of values.
It is recommended NOT to modify this argument; isntead, the \code{_params} argument should
be used for customization.}

\item{labs_params}{list. Additional parameters to pass to \code{ggplot2::labs()} to use in addition
to and/or override the parameters set in the \code{_base} parameter.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}

\item{theme_params}{list. Additional parameters to pass to \code{ggplot2::labs()}.
Should be used in the same manner as \code{labs_params} for customization.}
}
\value{
gg.
}
\description{
Visualize bigrams with dots sized according to frequency.
}
\details{
\code{compute_freqs_facet_at()} should NOT be called beforehand.
}
\seealso{
\url{https://buzzfeednews.github.io/2018-01-trump-twitter-wars/}
}
