% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{visualize_bigram_freqs_facet_at}
\alias{visualize_bigram_freqs_facet_at}
\alias{visualize_bigram_freqs_facet}
\title{Visualize bigrams}
\usage{
visualize_bigram_freqs_facet_at(data = NULL, token = NULL, facet = NULL,
  color = facet, color_value = "grey50", num_top = 3,
  title = "Most Frequently Used Pairs of Words", subtitle = NULL,
  caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext())

visualize_bigram_freqs_facet(data = NULL, token = NULL, facet = NULL,
  color = facet, color_value = "grey50", num_top = 3,
  title = "Most Frequently Used Pairs of Words", subtitle = NULL,
  caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext())
}
\arguments{
\item{data}{data.frame.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use as unit for counting.
Probably something like 'word' or 'bigram'.}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} to use for facetting.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{num_top}{numeric. Number of words to show. If between 0 and 1, then assumed to be a percentage.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}
}
\value{
gg.
}
\description{
Visualize bigrams with dots sized according to frequency.
}
\details{
\code{compute_freqs_facet_at()} should NOT be called beforehand.
}
\seealso{
\url{https://buzzfeednews.github.io/2018-01-trump-twitter-wars/}
}
