% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{visualize_bigram_freqs_multi}
\alias{visualize_bigram_freqs_multi}
\title{Visualize bigrams}
\usage{
visualize_bigram_freqs_multi(data = NULL, colname_word = "word",
  colname_multi = NULL, color = "grey50", num_top = 3,
  lab_title = "Most Frequently Used Pairs of Words",
  lab_subtitle = paste0("By ", stringr::str_to_title(colname_multi)),
  theme_base = temisc::theme_te_a())
}
\arguments{
\item{data}{data.frame.}

\item{colname_word}{character. Name of column in \code{data} to use for count. Default is provided}

\item{colname_multi}{character. Name of column in \code{data} to use for facetting.}

\item{color}{chracter. Hex value of color. Default is provided.}

\item{num_top}{numeric. Default is provided.}

\item{lab_title}{character. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}. Default is provided.}
}
\value{
gg
}
\description{
Visualize bigrams with dots sized according to frequency.
}
\details{
\code{compute_freqs_multi()} should NOT be called beforehand.
}
\seealso{
\url{https://buzzfeednews.github.io/2018-01-trump-twitter-wars/}
}
