% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_bigrams_freqs.R
\name{visualize_bigrams_freqs}
\alias{visualize_bigrams_freqs}
\title{Visualize bigrams}
\usage{
visualize_bigrams_freqs(data = NULL, colname_x = NULL,
  colname_word = "word", colname_freq = "freq", color = "grey50",
  num_top = 10, lab_title = "Most Frequently Used Pairs of Words",
  lab_subtitle = paste0("By ", stringr::str_to_title(colname_x)),
  theme_base = temisc::theme_te_b())
}
\arguments{
\item{data}{data.frame.}

\item{colname_x}{character. Name of column in \code{data} to use for x-axis.
Probably something like 'yyyy', 'mm', etc.}

\item{colname_word}{character. Name of column in \code{data} to use for bigram. Default: 'word'}

\item{colname_freq}{character. Name of column in \code{data} to use for ordering and sizing. Default: 'freq'}

\item{color}{chracter. Hex value of color. Default is provided.}

\item{num_top}{numeric.}

\item{lab_title}{character. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}. Default is provided.}
}
\value{
gg
}
\description{
Visualize bigrams with dots.
}
\details{
None.
}
\seealso{
\url{https://buzzfeednews.github.io/2018-01-trump-twitter-wars/}
}
