% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_bytime.R
\name{visualize_bytime}
\alias{visualize_bytime}
\title{Visualize by time period.}
\usage{
visualize_bytime(data, colname_x = NULL, geom = c("bar", "hist"),
  colname_color = NULL, color = "grey50", lab_subtitle = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{colname_x}{character. Probably something like 'timestamp', 'yyyy', 'mm', etc.}

\item{geom}{character. "bar" or "hist". "bar" is probably best for everything except Date objects.}

\item{colname_color}{character. Name of column in data.frame to use for color basis.}

\item{color}{chracter. Hex value of color. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.}
}
\value{
ggplot
}
\description{
Visualize time-based data over time.
}
\details{
Even though \code{colname_x} and \code{lab_subtitle} are not required, they
are sent to \code{NULL} by default in order to simplify the internal code.
}
