% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_bytime.R
\name{visualize_bytime_batched}
\alias{visualize_bytime_batched}
\title{Visualize over multiple time periods}
\usage{
visualize_bytime_batched(data = NULL, x_chars = c("timestamp", "yyyy", "mm",
  "wd", "hh"), geoms = c("hist", "bar", "bar", "bar", "bar"),
  labs_subtitle = c("", "By Year", "By Month", "By Day of Week", "By Hour"),
  colors = rep("grey50", 5), color_chars)
}
\arguments{
\item{data}{data.frame (single).}

\item{x_chars}{character (vector).}

\item{geoms}{character (vector).}

\item{labs_subtitle}{character (vector).}

\item{colors}{character (vector).}

\item{color_chars}{character (vector).}
}
\value{
ggplots
}
\description{
Visualize time-based data over time.
}
\details{
Calls \code{visualize_bytime()}.
}
