% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnts.R
\name{visualize_cnts_at}
\alias{visualize_cnts_at}
\alias{visualize_cnts}
\alias{visualize_cnts_multi_at}
\alias{visualize_cnts_multi}
\title{Visualize counts}
\usage{
visualize_cnts_at(data = NULL, word = "word", num_top = 10,
  color = NULL, color_value = "grey50", lab_title = "Count of Words",
  lab_subtitle = NULL, lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext())

visualize_cnts(data = NULL, word = "word", num_top = 10, color = NULL,
  color_value = "grey50", lab_title = "Count of Words",
  lab_subtitle = NULL, lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext())

visualize_cnts_multi_at(data = NULL, word = "word", num_top = 10,
  color = NULL, color_value = "grey50", lab_title = "Count of Words",
  lab_subtitle = NULL, lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext_facet(), multi)

visualize_cnts_multi(data = NULL, word = "word", num_top = 10,
  color = NULL, color_value = "grey50", lab_title = "Count of Words",
  lab_subtitle = NULL, lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext_facet(), multi)
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for words.
Probably something like 'word' or 'bigram'.}

\item{num_top}{numeric. Number of words to show. If between 0 and 1, then assumed to be a percentage.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function.}

\item{lab_title}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_subtitle}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_caption}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_x}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{theme_base}{\code{ggplot2} theme (e.g. as \code{ggplot2::theme_minimal()}.)
A custom \code{tetext} thems is supplied as a default.}

\item{multi}{character. Name of column to use as 'facetting' variable. Not used
if not specified.}
}
\value{
gg.

gg.
}
\description{
Visualize n-gram counts.
}
\details{
'multi' version is for facetting. 'y' value is calculated internally
with \code{dplyr::count()}.

Unfortunately, it seems difficult to combine this with \code{visualize_cnts_at()}
}
\seealso{
\url{https://github.com/dgrtwo/dgrtwo.github.com/blob/master/_R/2016-08-09-trump-data.Rmd}.
}
