% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnts.R
\name{visualize_cnts_at}
\alias{visualize_cnts_at}
\alias{visualize_cnts_multi_at}
\title{Visualize counts}
\usage{
visualize_cnts_at(data = NULL, word = "word", num_top = 10,
  color = NULL, color_value = "grey80", lab_title = "Count of Words",
  lab_subtitle = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = temisc::theme_te_a())

visualize_cnts_multi_at(data = NULL, word = "word", num_top = 10,
  color = NULL, color_value = "grey80", lab_title = "Count of Words",
  lab_subtitle = paste0("By ", stringr::str_to_title(multi)), lab_x = NULL,
  lab_y = NULL, theme_base = temisc::theme_te_a_facet(), multi)
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for words.
Probably something like 'word' or 'bigram'. Default is provided.}

\item{num_top}{numeric. Number of words to show. Default is provided.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function. Default is provided.}

\item{lab_title}{character. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{lab_x}{character. Default is provided.}

\item{lab_y}{character. Default is provided.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}.
Default is provided.}

\item{multi}{character. Name of column to use as 'facetting' variable. Not used
if not specified.}
}
\value{
gg.

gg.
}
\description{
Visualize n-gram counts.
}
\details{
'multi' version is for facetting. 'y' value is calculated internally
with \code{dplyr::count()}.

Unfortunately, it seems difficult to combine this with \code{visualize_cnts_at()}
}
\seealso{
\url{https://github.com/dgrtwo/dgrtwo.github.com/blob/master/_R/2016-08-09-trump-data.Rmd}.
}
