% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cnts_wordcloud.R
\name{visualize_cnts_wordcloud}
\alias{visualize_cnts_wordcloud}
\alias{visualize_cnts_wordcloud_multi}
\title{Visualize counts}
\usage{
visualize_cnts_wordcloud(data = NULL, colname_word = "word",
  colname_freq = "n", color = "grey50", num_top = 50,
  random_order = FALSE)

visualize_cnts_wordcloud_multi(data = NULL, ..., colname_multi = NULL,
  value_multi = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{colname_word}{character. Name of column in \code{data} to use for words. Default: 'word'.}

\item{colname_freq}{character. Name of column in \code{data} to use for frequencies. Default: 'n'.}

\item{color}{chracter. Hex value of color. Default is provided.}

\item{num_top}{numeric. Number of words to show.}

\item{random_order}{logical. Passed directly to \code{random.order} parameter
in \code{wordcloud::wordcloud()}}

\item{...}{dots. Parameters to pass to \code{visualize_cnts_wordcloud()}.}

\item{colname_multi}{character. Name of column in \code{data} to use to create
a separate wordcloud.}

\item{value_multi}{charater or numeric. Value by which to filter the column
specified by \code{colname_multi}.}
}
\value{
plot
}
\description{
Visualize n-gram counts with a word cloud.
}
\details{
'multi' version is for creating multiple plots simulatenously
(probably using a \code{purrr} 'map' function).
Fairly original function.
}
