% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{visualize_corrs_network_at}
\alias{visualize_corrs_network_at}
\alias{visualize_corrs_network}
\title{Visualize correlations}
\usage{
visualize_corrs_network_at(..., resize_points = TRUE, size_point = "n",
  add_point_labels = TRUE, color_point = "grey50", shape_point = 21,
  seed = 42, labs_base = labs_tetext(), labs_params = list(title =
  "Network of Pairwise Correlations"), theme_base = theme_tetext(void = TRUE),
  theme_params = list(legend.position = "none"))

visualize_corrs_network(..., token, feature)
}
\arguments{
\item{...}{dots. Parameters passed directly to \code{compute_corrs_at()}.}

\item{resize_points}{logical. Indicates whether or not to make size of points
correspond to count of words.}

\item{size_point}{bare for NSE; character for SE. Name of column in \code{data} to use for point sizing.}

\item{add_point_labels}{logical. Indicates whether or not to add labels to points.}

\item{color_point}{character. Hex value of color_value for points.}

\item{shape_point}{numeric.}

\item{seed}{numeric. Used to by \code{ggraph::ggraph}.}

\item{labs_base}{\code{ggplot2::labs()} function. Defaults to a pre-determined set of values.
It is recommended NOT to modify this argument; isntead, the \code{_params} argument should
be used for customization.}

\item{labs_params}{list. Additional parameters to pass to \code{ggplot2::labs()} to use in addition
to and/or override the parameters set in the \code{_base} parameter.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}

\item{theme_params}{list. Additional parameters to pass to \code{ggplot2::labs()}.
Should be used in the same manner as \code{labs_params} for customization.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use as unit for counting.
Probably something like 'word' or 'bigram'.}
}
\value{
gg.
}
\description{
Visualize correlations with a network
}
\seealso{
\url{https://www.tidytextmining.com/ngrams.html}.
\url{http://varianceexplained.org/r/seven-fav-packages/}.
}
