% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{visualize_freqs_facet_by2_at}
\alias{visualize_freqs_facet_by2_at}
\alias{visualize_freqs_facet_by2}
\title{Visualize token frequencies in pairs}
\usage{
visualize_freqs_facet_by2_at(data = NULL, token = NULL, facet = NULL,
  xy_nms, xy_grid, filter_facet = FALSE, facet_main = NULL,
  x_include = NULL, y_include = NULL, x_exclude = NULL,
  y_exclude = NULL, color = NULL, color_value = "grey50",
  add_labels = TRUE, title = "Relative Word Frequency", subtitle = NULL,
  caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext_facet(), scales = "free", ncol = 3,
  nrow = NULL, strip.position = "top")

visualize_freqs_facet_by2(data = NULL, token = NULL, facet = NULL, xy_nms,
  xy_grid, filter_facet = FALSE, facet_main = NULL, x_include = NULL,
  y_include = NULL, x_exclude = NULL, y_exclude = NULL, color = NULL,
  color_value = "grey50", add_labels = TRUE,
  title = "Relative Word Frequency", subtitle = NULL, caption = NULL,
  lab_x = NULL, lab_y = NULL, theme_base = theme_tetext_facet(),
  scales = "free", ncol = 3, nrow = NULL, strip.position = "top")
}
\arguments{
\item{data}{data.frame.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use for count. Default is provided}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} to use for facetting.}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{filter_facet}{logical. Whether or not to filter the \code{facet} values.}

\item{facet_main}{bare for NSE; character for SE. Name of single \code{facet} value to use as basis.
Not used if \code{filter_facet = FALSE}.}

\item{x_include, y_include, x_exclude, y_exclude}{character (vector).
\code{facet} values to include and exclude. Not used if \code{filter_facet = FALSE}.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{add_labels}{logical. Whether or not to add text labels (of the tokens).}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}
}
\value{
gg.
}
\description{
Visualize token frequenceis across pairs of \code{facet} values.
}
\details{
\code{compute_freqs_facet_by2_at()} should NOT be called beforehand.
}
