% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{visualize_freqs_multi_by2_at}
\alias{visualize_freqs_multi_by2_at}
\alias{visualize_freqs_multi_by2}
\title{Visualize n-gram frequencies in pairs}
\usage{
visualize_freqs_multi_by2_at(data = NULL, word = "word", multi = NULL,
  xy_nms, xy_grid, filter_multi = FALSE, multi_main = NULL,
  x_include = NULL, y_include = NULL, x_exclude = NULL,
  y_exclude = NULL, color = NULL, color_value = "grey50",
  add_labels = TRUE, lab_title = "Relative Word Frequency",
  lab_subtitle = NULL, lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext_facet(), facet_scales = "free",
  facet_ncol = 3, facet_nrow = NULL, facet_strip_position = "top")

visualize_freqs_multi_by2(data = NULL, word = "word", multi = NULL,
  xy_nms, xy_grid, filter_multi = FALSE, multi_main = NULL,
  x_include = NULL, y_include = NULL, x_exclude = NULL,
  y_exclude = NULL, color = NULL, color_value = "grey50",
  add_labels = TRUE, lab_title = "Relative Word Frequency",
  lab_subtitle = NULL, lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext_facet(), facet_scales = "free",
  facet_ncol = 3, facet_nrow = NULL, facet_strip_position = "top")
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for count. Default is provided}

\item{multi}{character. Name of column in \code{data} to use for facetting.}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{filter_multi}{logical. Whether or not to filter the \code{multi} values.}

\item{multi_main}{character. Name of single \code{multi} value to use as basis.
Not used if \code{filter_multi = FALSE}.}

\item{x_include, y_include, x_exclude, y_exclude}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function.}

\item{add_labels}{logical. Whether or not to add text labels (of the n-grams).}

\item{lab_title}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_subtitle}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_caption}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_x}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{theme_base}{\code{ggplot2} theme (e.g. as \code{ggplot2::theme_minimal()}.)
A custom \code{tetext} thems is supplied as a default.}

\item{facet_scales}{character. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_ncol}{numeric. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_nrow}{numeric. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_strip_position}{character. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}
}
\value{
gg.
}
\description{
Visualize n-gram frequenceis across pairs of \code{multi} values.
}
\details{
\code{compute_freqs_multi_by2_at()} should NOT be called beforehand.
}
