% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{visualize_hh_multi_at}
\alias{visualize_hh_multi_at}
\alias{visualize_hh_multi}
\title{Visualize hourly data}
\usage{
visualize_hh_multi_at(data = NULL, timebin = NULL, multi = NULL,
  color = multi, color_value = "grey50", lab_x = NULL, lab_y = NULL,
  lab_title = "Count Over Time", lab_subtitle = "By Time of Day",
  lab_caption = NULL, theme_base = theme_tetext())

visualize_hh_multi(data = NULL, timebin = NULL, multi = NULL,
  color = multi, color_value = "grey50", lab_x = NULL, lab_y = NULL,
  lab_title = "Count Over Time", lab_subtitle = "By Time of Day",
  lab_caption = NULL, theme_base = theme_tetext())
}
\arguments{
\item{data}{data.frame.}

\item{timebin}{character. Name of column in \code{data} to use for time. Probably likely something like 'hh' or 'hour'.}

\item{multi}{character. Name of column in \code{data} to use one of main axes, not for facetting.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function.}

\item{lab_x}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_title}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_subtitle}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_caption}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{theme_base}{\code{ggplot2} theme (e.g. as \code{ggplot2::theme_minimal()}.)
A custom \code{tetext} thems is supplied as a default.}
}
\description{
Visualize time-based data aggregated over hours.
}
\details{
This function is 'unique' because it calls \code{ggplot2::geom_violin()}.
}
