% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratio-by2.R
\name{visualize_logratios_facet_by2_at}
\alias{visualize_logratios_facet_by2_at}
\alias{visualize_logratios_facet_by2}
\title{Visualize token log ratios in pairs}
\usage{
visualize_logratios_facet_by2_at(data = NULL, token = NULL, facet = NULL,
  xy_nms, xy_grid, ..., filter_facet = FALSE, facet_main = NULL,
  filter_facet_base = default_filter_facet(facet_main),
  filter_facet_params = list(), color = NULL,
  lab_other = ifelse(!is.null(facet_main), paste0("Not ", facet_main),
  "Other"), num_top = 5, color_main = "grey50",
  scale_manual_base = default_scale_manual(generate_named_dual_colors(color_main,
  facet_main, lab_other)), scale_manual_params = list(),
  labs_base = default_labs(), labs_params = list(title =
  "Most Unique Words", y = "Log Odds Ratio"), theme_base = default_theme(),
  theme_params = list(legend.position = "bottom", axis.text.y =
  ggplot2::element_text(angle = 30, hjust = 1), panel.grid.major.y =
  ggplot2::element_blank()), facet_base = default_facet("name_xy"),
  facet_params = list())

visualize_logratios_facet_by2(..., token, color, facet)
}
\arguments{
\item{data}{data.frame. Data to apply a function over.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use for count.}

\item{facet}{character. Name of column in \code{data} to use for facetting.
It is set to NULL by default in the non-\code{facet} function even though it is not necessary.
If specified with the non-\code{facet} version of the function,
then the function acts just like the \code{_facet} version.}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{...}{dots. Additional parameters to pass to \code{compute_logratios_facet_by2_at()}.}

\item{filter_facet}{logical. Whether or not to filter the \code{facet} values.
\code{facet} values to include and exclude. Not used if \code{filter_facet = FALSE}.}

\item{facet_main}{of same type as \code{facet} values. Value of single 'main' facet.}

\item{filter_facet_base}{stuff. Work similarly to other \code{_base} arguments.}

\item{filter_facet_params}{list. Works similarly to other \code{_params} arguments.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{lab_other}{character. Name to give to 'opposing' factor of \code{facet_main}.}

\item{num_top}{numeric. Number of tokens to show for each \code{facet} pair.}

\item{color_main}{character. Opposite color is calculated for 'opposing' color.
It is advised to use this argument instead of setting \code{values} with the \code{scale_manual_params} argument.}

\item{scale_manual_base}{list. Parameters to pass to \code{ggplot2::scale_fill_manual()} or
\code{ggplot2::scale_color_manual()}, depending on the aesthetic mapping. A default \code{values}
argument is specified in the internal
function \code{default_scale_manual()},
so if overwriting explicitly, then \code{values} should be included in the list.}

\item{scale_manual_params}{list. Additional parameters to pass to \code{ggplot2::scale_fill_manual()}
or \code{ggplot2::scale_color_manual()} (e.g. \code{breaks}, \code{labels}, or \code{name}).
The default \code{values} specified by \code{scale_manual_base = default_scale_manual()}
can be overwritten a \code{values} argument in this list (or directly with a
\code{values} argument in the \code{scale_manual_base} list.}

\item{labs_base}{\code{ggplot2::labs()} function. Defaults to a pre-determined set of values.
It is recommended NOT to modify this argument; isntead, the \code{_params} argument should
be used for customization.}

\item{labs_params}{list. Additional parameters to pass to \code{ggplot2::labs()} to use in addition
to and/or override the parameters set in the \code{_base} parameter.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}

\item{theme_params}{list. Additional parameters to pass to \code{ggplot2::labs()}.
Should be used in the same manner as \code{labs_params} for customization.}

\item{facet_base}{\code{ggplot2::facet_wrap()} function. An internal function
sets defaults for \code{facets}, \code{scales}, \code{ncol}, \code{nrow}, and \code{strip.position}
As with the \code{labs_base} and \code{theme_base} arguments, it is NOT recommended to set
this argument explicitly; instead, the accompanying \code{_params} argument should be used.}

\item{facet_params}{list. Additional parameters to pass to \code{ggplot2::facet_wrap()}. Parameters
passed to this argument override thos in the \code{_base} argument.}
}
\value{
gg.
}
\description{
Visualize token log ratios across pairs of \code{facet} values.
}
\details{
\code{compute_logratios_facet_by2_at()} should NOT be called beforehand.
}
\seealso{
\url{https://www.tidytextmining.com/twitter.html#comparing-token-usage}.
\url{http://varianceexplained.org/r/trump-tweets/}.
}
