% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratio.R
\name{visualize_logratios_multi_by2_at}
\alias{visualize_logratios_multi_by2_at}
\alias{visualize_logratios_multi_by2}
\title{Visualize n-gram log ratios in pairs}
\usage{
visualize_logratios_multi_by2_at(data = NULL, word = "word",
  xy_nms = NULL, xy_grid = NULL, multi = NULL, ..., filter_multi = TRUE,
  x_include = NULL, y_include = NULL, x_exclude = NULL,
  y_exclude = NULL, multi_main = NULL, color = NULL,
  color_value = c("grey50", "grey80"), num_top = 3,
  lab_title = "Most Unique Words", lab_subtitle = paste0("By ",
  stringr::str_to_title(multi)), lab_x = NULL, lab_y = "Log Odds Ratio",
  theme_base = theme_tetext_facet(), facet_scales = "free",
  facet_ncol = 3, facet_nrow = NULL, facet_strip_position = "top")

visualize_logratios_multi_by2(data = NULL, word = "word", xy_nms = NULL,
  xy_grid = NULL, multi = NULL, ..., filter_multi = TRUE,
  x_include = NULL, y_include = NULL, x_exclude = NULL,
  y_exclude = NULL, multi_main = NULL, color = NULL,
  color_value = c("grey50", "grey80"), num_top = 3,
  lab_title = "Most Unique Words", lab_subtitle = paste0("By ",
  stringr::str_to_title(multi)), lab_x = NULL, lab_y = "Log Odds Ratio",
  theme_base = theme_tetext_facet(), facet_scales = "free",
  facet_ncol = 3, facet_nrow = NULL, facet_strip_position = "top")
}
\arguments{
\item{data}{data.frame. Data to apply a function over.}

\item{word}{character. Name of column in \code{data} to use for count. Default is provided}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{multi}{character. Name of column in \code{data} corresponding to group
by which count is made.}

\item{...}{Additional parameters to pass to \code{compute_logratios_multi_by2_at()}.}

\item{filter_multi}{logical. Whether or not to filter the \code{multi} values.}

\item{x_include}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{y_include}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{x_exclude}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{y_exclude}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{multi_main}{character. Name of single \code{multi} value to use as
basis.
Essentially equivalent to how \code{x_include} should be used. Not used if \code{filter_multi = FALSE}.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function.}

\item{num_top}{numeric. Number of n-grams to show for each \code{multi} pair.}

\item{lab_title}{character.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{lab_x}{character.}

\item{lab_y}{character.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}.}

\item{facet_scales}{character. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_ncol}{numeric. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_nrow}{numeric. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_strip_position}{character. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}
}
\value{
gg.
}
\description{
Visualize n-grams log ratios across pairs of \code{multi} values.
}
\details{
\code{compute_logratios_multi_by2_at()} should NOT be called beforehand.
}
