% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{visualize_tfidf_at}
\alias{visualize_tfidf_at}
\title{Visualize TF-IDF}
\usage{
visualize_tfidf_at(data = NULL, word = "word", doc = "document",
  multi = doc, num_top = 10, color = multi, color_value = "grey80",
  lab_title = "Highest TF-IDF Words", lab_subtitle = paste0("By ",
  stringr::str_to_title(multi)), theme_base = temisc::theme_te_a_facet())
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for words.
Probably something like 'word' or 'bigram'. Default is provided.}

\item{doc}{character. Name of column in \code{data} to use
as \code{document} in \code{tidytext::bind_tf_idf()}. Default is provided.}

\item{multi}{character. Name of column in \code{data} to use for facetting.}

\item{num_top}{numeric. Number of words to show. Default is provided.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function. Default is provided.}

\item{lab_title}{character. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}.
Default is provided.}
}
\value{
gg.
}
\description{
Visualize term-frequency, inverse-document-frequency.
}
\details{
Calls \code{compute_tfidf_at()} internally.
Cannot pass dots to \code{compute_tfidf_at()} internally because parameters
are used in subsequent processing.
}
\seealso{
\url{https://www.tidytextmining.com/tfidf.html}.
\url{https://juliasilge.com/blog/sherlock-holmes-stm/}
}
