% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{visualize_tfidf_multi}
\alias{visualize_tfidf_multi}
\title{Visualize TF-IDF}
\usage{
visualize_tfidf_multi(data = NULL, colname_word = "word",
  colname_tfidf = "tfidf", colname_multi = "document", num_top = 10,
  colname_color = colname_multi, color = "grey50",
  lab_title = "Highest TF-IDF Words", lab_subtitle = paste0("By ",
  stringr::str_to_title(colname_multi)),
  theme_base = temisc::theme_te_a_facet())
}
\arguments{
\item{data}{data.frame.}

\item{colname_word}{character. Name of column in \code{data} to use for x-axis.
Probably something like 'word' or 'bigram'. Default is provided.}

\item{colname_tfidf}{character Name of column in \code{data} to use for y-axis. Default is provided}

\item{colname_multi}{character. Name of column to use as 'facetting' variable. Not used
if not specified.}

\item{num_top}{numeric. Default is provided.}

\item{colname_color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color}{chracter. Hex value of color. Default is provided.}

\item{lab_title}{character. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}. Default is provided.}
}
\description{
Visualize term-frequency, inverse-document-frequency.
}
\details{
Does not call \code{compute_tfidf()} internally.
}
