% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_time.R
\name{visualize_time_batched}
\alias{visualize_time_batched}
\title{Visualize over multiple time periods}
\usage{
visualize_time_batched(data = NULL, colnames_x = c("timestamp", "yyyy",
  "mm", "wd", "hh"), geoms = c("hist", "bar", "bar", "bar", "bar"),
  labs_subtitle = c("", "By Year", "By Month", "By Day of Week", "By Hour"),
  colors = rep("grey50", 5), color_chars)
}
\arguments{
\item{data}{data.frame (single).}

\item{colnames_x}{character (vector).}

\item{geoms}{character (vector).}

\item{labs_subtitle}{character (vector).}

\item{colors}{character (vector).}

\item{color_chars}{character (vector).}
}
\value{
gg
}
\description{
Visualize time-based data over time.
}
\details{
Calls \code{visualize_time()}.
}
