% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{compute_change_at}
\alias{compute_change_at}
\alias{compute_change}
\title{Compute change}
\usage{
compute_change_at(data = NULL, word = "word", timebin = NULL,
  bin = TRUE, timefloor = NULL, top_pct = 0.25, only_signif = FALSE,
  signif_cutoff = 0.1, return_models = TRUE, return_data = FALSE,
  return_both = FALSE)

compute_change(data = NULL, word = "word", timebin = NULL, bin = TRUE,
  timefloor = NULL, top_pct = 0.25, only_signif = FALSE,
  signif_cutoff = 0.1, return_models = TRUE, return_data = FALSE,
  return_both = FALSE)
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} corresponding to n-gram.}

\item{timebin}{character. Name of column in \code{data} specifying temporal period
to use to compute change.}

\item{bin}{logical. Whether or not to call \code{lubridate::floor_date()} to truncate \code{timebin}.}

\item{timefloor}{character. Name of column passed directly to \code{unit} parameter of
\code{lubridate::floor_date()} if \code{bin = FALSE}.}

\item{top_pct}{numeric. Number between 0 and 1. Useful primarily to limit the number of models
that need to be computed and to reduce 'nose' regarding what is deemed significant.}

\item{only_signif}{logical. Whether or not to return rows with a significant p-value.}

\item{signif_cutoff}{numeric. Number between 0 and 1. Value to use as 'maximum' threshhld for significance.
Only used if \code{only_signif = TRUE}.}

\item{return_models}{logical. Whether to return just the models. This is probably the preferred
option when calling \code{compute_change_at()} directly.}

\item{return_data}{logical. Whether to 'bytime' data which is used as
the \code{data} parameter in \code{stats::glm()} for creating models.
Needed when using \code{visualize_change_at()}.}

\item{return_both}{logical. Set to \code{TRUE} when \code{visualize_change_at()}.}
}
\value{
data.frame.
}
\description{
Compute change in n-gram usage over time.
}
\details{
None.
}
\seealso{
\url{https://www.tidytextmining.com/twitter.html#changes-in-word-use}
}
