% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{compute_change}
\alias{compute_change}
\title{Compute change}
\usage{
compute_change(data = NULL, colname_timebin = NULL, colname_word = "word",
  timefloor = c("second", "year", "hour", "day", "week", "year"),
  top_pct = 0.05)
}
\arguments{
\item{data}{data.frame.}

\item{colname_timebin}{character. Name of column in \code{data} specifying temporal period
to use to compute change.}

\item{colname_word}{character. Name of column in \code{data} corresponding to n-gram.}

\item{timefloor}{character. Passed directly to \code{lubridate::floor_date()} as\code{unit} parameter.}

\item{top_pct}{numeric. Number between 0 and 1. Default is provided.}
}
\value{
data.frame
}
\description{
Compute change in word usage over time.
}
\details{
None.
}
\seealso{
https://www.tidytextmining.com/twitter.html#changes-in-word-use
}
