% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{compute_corrs}
\alias{compute_corrs}
\title{Compute correlations}
\usage{
compute_corrs(data = NULL, colname_word = "word", colname_feature = NULL,
  num_top_ngrams = 50, num_top_corrs = 50, return_corrs = TRUE,
  return_words = FALSE, return_both = FALSE)
}
\arguments{
\item{data}{data.frame.}

\item{colname_word}{character. Name of column in \code{data} to use as \code{item}
in \code{widyr::pairwise_cor()}. Default is provided.}

\item{colname_feature}{character. Name of column in \code{data} to use as \code{feature}
in \code{widyr::pairwise_cor()}. No default is provided}

\item{num_top_ngrams}{numeric. Useful primarily to prevent \code{widyr::pairwise_cor()}
from hanging up. Default is provided.}

\item{num_top_corrs}{numeric. Useful primarily to limit input to a network visualation.
does not hang up. Default is provided.}

\item{return_corrs}{logical. Whether to return just the correlations and word pairs,
but not the word counts. This is the default option.}

\item{return_words}{logical. Whether to return the counts used to compute
the correlations. Useful when creating a network visualization.}

\item{return_both}{logical. Useful when creating a network visualization so that
words can be used as nodes and correlations can be used weights.}
}
\value{
data.frame.
}
\description{
Compute correlations among pairs of words.
}
\details{
Call \code{widyr::pairwise_cor()} internally.
Is called by \code{visualize_corrs_network()}, so there is no need to call this directly.
}
\seealso{
\url{http://varianceexplained.org/r/seven-fav-packages/}
}
