% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{compute_corrs_at}
\alias{compute_corrs_at}
\alias{compute_corrs}
\title{Compute correlations}
\usage{
compute_corrs_at(data = NULL, token = NULL, feature = NULL,
  num_top_ngrams = 50, num_top_corrs = 50, adjust = FALSE,
  return_corrs = TRUE, return_words = FALSE, return_both = FALSE)

compute_corrs(..., token, feature)
}
\arguments{
\item{data}{data.frame.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use as \code{item}
in \code{widyr::pairwise_cor()}.}

\item{feature}{bare for NSE; character for SE. Name of column in \code{data} to use as \code{feature}
in \code{widyr::pairwise_cor()}.}

\item{num_top_ngrams}{numeric. Useful primarily to prevent \code{widyr::pairwise_cor()}
from hanging up. If between 0 and 1, then interpreted as a percentage.}

\item{num_top_corrs}{numeric. Useful primarily to limit input to a network visualation.
does not hang up. If between 0 and 1, then interpreted as a percentage.}

\item{adjust}{logical. Indicates whether or not to scale \code{num_top_ngrams} and
\code{num_top_corrs} (if they are between 0 and 1) to the count of unique words.
Is is likely that the input percentage values are specified with the number of words in mind.}

\item{return_corrs}{logical. Whether to return just the correlations and token pairs,
but not the token counts. This is the default option.}

\item{return_words}{logical. Whether to return the counts used to compute
the correlations. Useful when creating a network visualization.}

\item{return_both}{logical. Useful when creating a network visualization so that
words can be used as nodes and correlations can be used weights.}

\item{...}{dots. Parameters to pass directly to \code{visualize_time()}.}
}
\value{
data.frame.
}
\description{
Compute correlations among pairs of words.
}
\details{
Calls \code{widyr::pairwise_cor()} internally.
}
\seealso{
\url{http://varianceexplained.org/r/seven-fav-packages/}.
}
