% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_freqs.R
\name{compute_freqs}
\alias{compute_freqs}
\alias{compute_freqs_multi}
\title{Compute n-gram frequency}
\usage{
compute_freqs(data = NULL, colname_cnt = "word")

compute_freqs_multi(data = NULL, colname_cnt = "word",
  colname_multi = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{colname_cnt}{character. Name of column in \code{data} to use for count. Default: 'word'.}

\item{colname_multi}{character. Name of column in \code{data} corresponding to group
by which count is made.}
}
\value{
data.frame.
}
\description{
Compute the frequency of n-grams.
}
\details{
Inspired by \href{https://www.tidytextmining.com/}{\emph{Text Mining with R}}.
}
