% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{compute_freqs_facet_at}
\alias{compute_freqs_facet_at}
\alias{compute_freqs_facet}
\alias{compute_freqs_at}
\alias{compute_freqs}
\alias{compute_freqs_facet_wide_at}
\title{Compute token frequency}
\usage{
compute_freqs_facet_at(data = NULL, token = NULL, facet = NULL)

compute_freqs_facet(..., token, facet)

compute_freqs_at(data = NULL, token = NULL, facet = NULL)

compute_freqs(..., token, facet)

compute_freqs_facet_wide_at(..., token = NULL, facet = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use for count.}

\item{facet}{character. Name of column in \code{data} to use for facetting.
It is set to NULL by default in the non-\code{facet} function even though it is not necessary.
If specified with the non-\code{facet} version of the function,
then the function acts just like the \code{_facet} version.}

\item{...}{dots. Parameters to pass directly to \code{visualize_time()}.}
}
\value{
data.frame.

data.frame.
}
\description{
Compute the frequency of tokens.

Manipulates output from \code{compute_freqs_facet_at()}
for \code{compute_freqs_mult_by2} function.
}
\details{
None.

To be used by \code{_by2} function only.
}
