% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{compute_freqs_at}
\alias{compute_freqs_at}
\alias{compute_freqs}
\alias{compute_freqs_multi_at}
\alias{compute_freqs_multi}
\title{Compute n-gram frequency}
\usage{
compute_freqs_at(data = NULL, word = "word")

compute_freqs(data = NULL, word = "word")

compute_freqs_multi_at(data = NULL, word = "word", multi = NULL)

compute_freqs_multi(data = NULL, word = "word", multi = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for count. Default is provided}

\item{multi}{character. Name of column in \code{data} corresponding to group
by which count is made.}
}
\value{
data.frame.

data.frame.
}
\description{
Compute the frequency of n-grams.
}
\details{
Inspired by \href{https://www.tidytextmining.com/}{\emph{Text Mining with R}}.
}
