% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{compute_tfidf}
\alias{compute_tfidf}
\title{Compute TF-IDF}
\usage{
compute_tfidf(data = NULL, colname_word = "word",
  colname_doc = "document")
}
\arguments{
\item{data}{data.frame.}

\item{colname_word}{character. Name of column in \code{data} to use
as \code{term} in \code{tidytext::bind_tf_idf()}. Default is provided.}

\item{colname_doc}{character. Name of column in \code{data} to use
as \code{document} in \code{tidytext::bind_tf_idf()}. Default is provided.}
}
\description{
Compute term-frequency, inverse-document-frequency.
}
\details{
Calls \code{tidytext::bind_tf_idf()} internally.
}
