% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default_theme}
\alias{default_theme}
\title{Default theme parameters}
\usage{
default_theme(base_family = "Arial Narrow", base_size = 11.5,
  plot_title_size = 18, subtitle_size = 12, void = FALSE, ...)
}
\arguments{
\item{base_family}{character.}

\item{base_size, plot_title_size, subtitle_size}{numeric.}

\item{void}{logical. Whether or not to modifly theme to mimic \code{ggplot2::theme_void()}/}

\item{...}{dots. Additional parameters passed to \code{ggplot2::theme()}.}
}
\value{
function.
}
\description{
A custom \code{ggplot2} theme.
}
\details{
This theme is based on \code{ggplot2::theme_minimal()}.
}
