% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_tetext}
\alias{theme_tetext}
\alias{theme_tetext_dx}
\alias{theme_tetext_facet}
\alias{theme_tetext_facet_dx}
\title{Custom theme}
\usage{
theme_tetext(base_family = "", base_size = 11, plot_title_size = 16,
  subtitle_size = 12, void = FALSE, ...)

theme_tetext_dx(...)

theme_tetext_facet(...)

theme_tetext_facet_dx(...)
}
\arguments{
\item{base_family}{character.}

\item{base_size, plot_title_size, subtitle_size}{numeric.}

\item{void}{logical. Whether or not to modifly theme to mimic \code{ggplot2::theme_void()}/}

\item{...}{dots. Additional parameters passed to \code{ggplot2::theme()}.}
}
\description{
A custom \code{ggplot2} theme (that is similar to \code{ggplot2::theme_minimal()}.
}
\details{
This theme is 'copy-pasted' of theme from \code{teplot} package.
}
\seealso{
\url{https://github.com/hrbrmstr/hrbrthemes/blob/master/R/theme-ipsum.r}
}
