% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{visualize_bigram_freqs_multi_at}
\alias{visualize_bigram_freqs_multi_at}
\alias{visualize_bigram_freqs_multi}
\title{Visualize bigrams}
\usage{
visualize_bigram_freqs_multi_at(data = NULL, word = "word", multi = NULL,
  color = multi, color_value = "grey50", num_top = 3,
  lab_title = "Most Frequently Used Pairs of Words", lab_subtitle = NULL,
  lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext())

visualize_bigram_freqs_multi(data = NULL, word = "word", multi = NULL,
  color = multi, color_value = "grey50", num_top = 3,
  lab_title = "Most Frequently Used Pairs of Words", lab_subtitle = NULL,
  lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext())
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for words.
Probably something like 'word' or 'bigram'.}

\item{multi}{character. Name of column in \code{data} to use for facetting.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function.}

\item{num_top}{numeric. Number of words to show. If between 0 and 1, then assumed to be a percentage.}

\item{lab_title}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_subtitle}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_caption}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_x}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{theme_base}{\code{ggplot2} theme (e.g. as \code{ggplot2::theme_minimal()}.)
A custom \code{tetext} thems is supplied as a default.}
}
\value{
gg.
}
\description{
Visualize bigrams with dots sized according to frequency.
}
\details{
\code{compute_freqs_multi_at()} should NOT be called beforehand.
}
\seealso{
\url{https://buzzfeednews.github.io/2018-01-trump-twitter-wars/}
}
