% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{visualize_bigram_freqs_multi_at}
\alias{visualize_bigram_freqs_multi_at}
\title{Visualize bigrams}
\usage{
visualize_bigram_freqs_multi_at(data = NULL, word = "word", multi = NULL,
  color = multi, color_value = "grey80", num_top = 3,
  lab_title = "Most Frequently Used Pairs of Words",
  lab_subtitle = paste0("By ", stringr::str_to_title(multi)), lab_x = NULL,
  lab_y = NULL, theme_base = temisc::theme_te_a())
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for words.
Probably something like 'word' or 'bigram'. Default is provided.}

\item{multi}{character. Name of column in \code{data} to use for facetting.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function. Default is provided.}

\item{num_top}{numeric. Number of words to show. Default is provided.}

\item{lab_title}{character. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{lab_x}{character. Default is provided.}

\item{lab_y}{character. Default is provided.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}.
Default is provided.}
}
\value{
gg.
}
\description{
Visualize bigrams with dots sized according to frequency.
}
\details{
\code{compute_freqs_multi_at()} should NOT be called beforehand.
}
\seealso{
\url{https://buzzfeednews.github.io/2018-01-trump-twitter-wars/}
}
