% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{visualize_change_at}
\alias{visualize_change_at}
\alias{visualize_change}
\title{Visualize change}
\usage{
visualize_change_at(..., token = NULL, num_top = 5, color = token,
  add_labels = FALSE, scale_manual_base = default_scale_manual(values =
  rep("grey50", num_top)), scale_manual_params = list(),
  labs_base = default_labs(), labs_params = list(title =
  "Tokens with Most Significant Change in Frequency", caption =
  paste0("Statistical significance is determined by a logistical model\\n",
  "estimating token appearance in a given time period.")),
  theme_base = default_theme(panel.grid.major.x = ggplot2::element_blank()),
  theme_params = list(legend.position = ifelse(add_labels, "none", "bottom")))

visualize_change(..., token, timebin, timefloor, color)
}
\arguments{
\item{...}{dots. Parameters passed to \code{compute_change_at()}.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use as unit for counting.
Probably something like 'word' or 'bigram'.}

\item{num_top}{numeric. Number of words to show. If between 0 and 1, then assumed to be a percentage.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{add_labels}{logical. Whether or not to add labels to the lines.}

\item{scale_manual_base}{list. Parameters to pass to \code{ggplot2::scale_fill_manual()} or
\code{ggplot2::scale_color_manual()}, depending on the aesthetic mapping. A default \code{values}
argument is specified in the internal
function \code{default_scale_manual()},
so if overwriting explicitly, then \code{values} should be included in the list.}

\item{scale_manual_params}{list. Additional parameters to pass to \code{ggplot2::scale_fill_manual()}
or \code{ggplot2::scale_color_manual()} (e.g. \code{breaks}, \code{labels}, or \code{name}).
The default \code{values} specified by \code{scale_manual_base = default_scale_manual()}
can be overwritten a \code{values} argument in this list (or directly with a
\code{values} argument in the \code{scale_manual_base} list.}

\item{labs_base}{\code{ggplot2::labs()} function. Defaults to a pre-determined set of values.
It is recommended NOT to modify this argument; isntead, the \code{_params} argument should
be used for customization.}

\item{labs_params}{list. Additional parameters to pass to \code{ggplot2::labs()} to use in addition
to and/or override the parameters set in the \code{_base} parameter.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}

\item{theme_params}{list. Additional parameters to pass to \code{ggplot2::labs()}.
Should be used in the same manner as \code{labs_params} for customization.}

\item{timebin}{bare for NSE; character for SE. Name of column in \code{data} to use for time axis.}

\item{timefloor}{character. Name of column passed directly to \code{unit} parameter of
\code{lubridate::floor_date()} if \code{bin = FALSE}.}
}
\value{
gg.
}
\description{
Visualize the words that have changed the most across \code{timebin}
}
\details{
Calls \code{compute_change()} internally. Works similarly to \code{visualize_corrs}.
(Two data sets are returned from the list that is returned by the \code{compute_} function.)
}
