% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cnts.R
\name{visualize_cnts}
\alias{visualize_cnts}
\alias{visualize_cnts_multi}
\title{Visualize counts}
\usage{
visualize_cnts(data = NULL, colname_x = "word", colname_y = "n",
  num_top = 10, colname_color = NULL, color = "grey50",
  lab_title = "Count of Words", lab_subtitle = NULL,
  theme_base = temisc::theme_te_b())

visualize_cnts_multi(data = NULL, colname_x = "word", colname_y = "n",
  num_top = 10, colname_color = NULL, color = "grey50",
  lab_title = "Count of Words", lab_subtitle = paste0("By ",
  stringr::str_to_title(colname_multi)),
  theme_base = temisc::theme_te_b_facet(), colname_multi)
}
\arguments{
\item{data}{data.frame.}

\item{colname_x}{character. Name of column in \code{data} to use for x-axis.
Probably something like 'word' or 'bigram'.}

\item{colname_y}{character.}

\item{num_top}{numeric. Default: 10.}

\item{colname_color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color}{chracter. Hex value of color. Default is provided.}

\item{lab_title}{character. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}. Default is provided.}

\item{colname_multi}{character. Name of column to use as 'facetting' variable. Not used
if not specified.}
}
\value{
gg
}
\description{
Visualize n-gram counts.
}
\details{
'multi' version is for facetting.
}
\seealso{
\url{https://github.com/dgrtwo/dgrtwo.github.com/blob/master/_R/2016-08-09-trump-data.Rmd}.
}
