% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnts.R
\name{visualize_cnts_wordcloud_at}
\alias{visualize_cnts_wordcloud_at}
\alias{visualize_cnts_wordcloud}
\alias{visualize_cnts_wordcloud_multi_at}
\alias{visualize_cnts_wordcloud_multi}
\title{Visualize counts}
\usage{
visualize_cnts_wordcloud_at(data = NULL, word = "word",
  color_value = "grey50", num_top = 50, random_order = FALSE)

visualize_cnts_wordcloud(data = NULL, word = "word",
  color_value = "grey50", num_top = 50, random_order = FALSE)

visualize_cnts_wordcloud_multi_at(data = NULL, ..., multi = NULL,
  all = FALSE, value_multi = NULL, num_par_row, num_par_col)

visualize_cnts_wordcloud_multi(data = NULL, ..., multi = NULL,
  all = FALSE, value_multi = NULL, num_par_row, num_par_col)
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for words.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function.}

\item{num_top}{numeric. Number of words to show.}

\item{random_order}{logical. Passed directly to \code{random.order} parameter
in \code{wordcloud::wordcloud()}}

\item{...}{dots. Parameters to pass to \code{visualize_cnts_wordcloud_at()}.}

\item{multi}{character. Name of column in \code{data} to use to create
a separate wordcloud.}

\item{all}{logical. Whether or not to create a wordcloud for each value of \code{multi}.}

\item{value_multi}{charater or numeric. Value by which to filter the column
specified by \code{multi}. Ignored if \code{all = TRUE}.}

\item{num_par_row, num_par_col}{numeric. Values to pass to \code{mfrow} in \code{par}.
Defaults are provided internally based on value of \code{all} and number of \code{multi} values.}
}
\value{
plot.

plots.
}
\description{
Visualize n-gram counts with a word cloud.
}
\details{
'multi' version is for creating multiple plots simulatenously
(probably using a \code{purrr} 'map' function).
Fairly original function.
}
