% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnts.R
\name{visualize_cnts_wordcloud_at}
\alias{visualize_cnts_wordcloud_at}
\alias{visualize_cnts_wordcloud}
\alias{visualize_cnts_wordcloud_facet_at}
\alias{visualize_cnts_wordcloud_facet}
\title{Visualize counts}
\usage{
visualize_cnts_wordcloud_at(data = NULL, token = NULL,
  wordcloud_base = wordcloud_tetext(data, token), wordcloud_params)

visualize_cnts_wordcloud(..., token)

visualize_cnts_wordcloud_facet_at(data = NULL, ..., facet = NULL,
  value_facet = NULL)

visualize_cnts_wordcloud_facet(..., token, facet)
}
\arguments{
\item{data}{data.frame.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data}
to use for \code{words} argument in \code{wordcloud::wordcloud()}.}

\item{...}{dots. Parameters to pass to \code{visualize_cnts_wordcloud_at()}.}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} to use to create
a separate wordcloud.}

\item{value_facet}{charater or numeric. Value by which to filter the column
specified by \code{facet}. Ignored if \code{all = TRUE}.
Defaults are provided internally based on value of \code{all} and number of \code{facet} values.}

\item{colors}{character. Passed directly to \code{colors} argument in \code{wordcloud::wordcloud()}.}

\item{num_top}{numeric. Number of words to show. Passed directly to \code{max.words} argument}

\item{random.order}{logical. Passed directly to \code{random.order} argument
in \code{wordcloud::wordcloud()}.}
}
\value{
plot.

plots.
}
\description{
Visualize token counts with a token cloud.
}
\details{
\code{facet} version is for creating multiple plots simulatenously
(probably using a \code{purrr} 'map' function).
Calls \code{wordcloud::wordcloud()}.
}
