% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{visualize_corrs_network_at}
\alias{visualize_corrs_network_at}
\alias{visualize_corrs_network}
\title{Visualize correlations}
\usage{
visualize_corrs_network_at(..., resize_points = TRUE, size_point = "n",
  add_point_labels = TRUE, color_point = "grey50", shape_point = 21,
  seed = 42, lab_title = "Network of Pairwise Correlations",
  lab_subtitle = NULL, lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext(void = TRUE))

visualize_corrs_network(..., resize_points = TRUE, size_point = "n",
  add_point_labels = TRUE, color_point = "grey50", shape_point = 21,
  seed = 42, lab_title = "Network of Pairwise Correlations",
  lab_subtitle = NULL, lab_caption = NULL, lab_x = NULL, lab_y = NULL,
  theme_base = theme_tetext(void = TRUE))
}
\arguments{
\item{...}{dots. Parameters passed directly to \code{compute_corrs_at()}.}

\item{resize_points}{logical. Indicates whether or not to make size of points
correspond to count of words.}

\item{size_point}{character. Name of column in \code{data} to use for point sizing.}

\item{add_point_labels}{logical. Indicates whether or not to add labels to points.}

\item{color_point}{character. Hex value of color_value for points.}

\item{shape_point}{numeric.}

\item{seed}{numeric. Used to by \code{ggraph::ggraph}.}

\item{lab_title}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_subtitle}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_caption}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_x}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{theme_base}{\code{ggplot2} theme (e.g. as \code{ggplot2::theme_minimal()}.)
A custom \code{tetext} thems is supplied as a default.}
}
\value{
gg.
}
\description{
Visualize correlations with a network
}
\seealso{
\url{https://www.tidytextmining.com/ngrams.html}.
\url{http://varianceexplained.org/r/seven-fav-packages/}.
}
