% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{visualize_hh_multi_at}
\alias{visualize_hh_multi_at}
\alias{visualize_hh_multi}
\title{Visualize hourly data}
\usage{
visualize_hh_multi_at(data = NULL, timebin = NULL, multi = NULL,
  color = multi, color_value = "grey80", lab_title = "Count Over Time",
  lab_subtitle = "By Time of Day", lab_x = NULL, lab_y = NULL,
  theme_base = temisc::theme_te_a())

visualize_hh_multi(data = NULL, timebin = NULL, multi = NULL,
  color = multi, color_value = "grey80", lab_title = "Count Over Time",
  lab_subtitle = "By Time of Day", lab_x = NULL, lab_y = NULL,
  theme_base = temisc::theme_te_a())
}
\arguments{
\item{data}{data.frame.}

\item{timebin}{character. Name of column in \code{data} to use for time. Probably likely something like 'hh' or 'hour'.}

\item{multi}{character. Name of column in \code{data} to use one of main axes, not for facetting.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function. Default is provided.}

\item{lab_title}{character. Default is provided.}

\item{lab_subtitle}{character. Probably something like 'By Year', 'By Month', etc.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{lab_x}{character. Default is provided.}

\item{lab_y}{character. Default is provided.}

\item{theme_base}{\code{ggplot2} theme, such as \code{ggplot2::theme_minimal()}.
Default is provided.}
}
\description{
Visualize time-based data aggregated over hours.
}
\details{
This function is 'unique' because it calls \code{ggplot2::geom_violin()}.
}
