% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratio.R
\name{visualize_logratios_multi_by2_at}
\alias{visualize_logratios_multi_by2_at}
\alias{visualize_logratios_multi_by2}
\title{Visualize n-gram log ratios in pairs}
\usage{
visualize_logratios_multi_by2_at(data = NULL, word = "word", multi = NULL,
  xy_nms, xy_grid, ..., filter_multi = TRUE, multi_main = NULL,
  x_include = NULL, y_include = NULL, x_exclude = NULL,
  y_exclude = NULL, color = NULL, color_value = c("grey50", "grey80"),
  num_top = 3, flip_axes = FALSE, lab_other = "other",
  lab_title = "Most Unique Words", lab_subtitle = NULL,
  lab_caption = NULL, lab_x = NULL, lab_y = "Log Odds Ratio",
  theme_base = theme_tetext_facet(), facet_scales = "free",
  facet_ncol = 3, facet_nrow = NULL, facet_strip_position = "right")

visualize_logratios_multi_by2(data = NULL, word = "word", multi = NULL,
  xy_nms, xy_grid, ..., filter_multi = TRUE, multi_main = NULL,
  x_include = NULL, y_include = NULL, x_exclude = NULL,
  y_exclude = NULL, color = NULL, color_value = c("grey50", "grey80"),
  num_top = 3, flip_axes = FALSE, lab_other = "other",
  lab_title = "Most Unique Words", lab_subtitle = NULL,
  lab_caption = NULL, lab_x = NULL, lab_y = "Log Odds Ratio",
  theme_base = theme_tetext_facet(), facet_scales = "free",
  facet_ncol = 3, facet_nrow = NULL, facet_strip_position = "right")
}
\arguments{
\item{data}{data.frame. Data to apply a function over.}

\item{word}{character. Name of column in \code{data} to use for count. Default is provided}

\item{multi}{character. Name of column in \code{data} corresponding to group
by which count is made.}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{...}{dots. Additional parameters to pass to \code{compute_logratios_multi_by2_at()}.}

\item{filter_multi}{logical. Whether or not to filter the \code{multi} values.}

\item{multi_main}{character. Name of single \code{multi} value to use as basis.
Not used if \code{filter_multi = FALSE}.}

\item{x_include}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{y_include}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{x_exclude}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{y_exclude}{character (vector).
\code{multi} values to include and exclude. Not used if \code{filter_multi = FALSE}.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Should be a vector of length two (for dual colors)
if \code{filter_multi = TRUE} and \code{!is.null(multi_main) & length(multi_main) == 1}.}

\item{num_top}{numeric. Number of n-grams to show for each \code{multi} pair.}

\item{flip_axes}{logical. Whether or not to call \code{ggplot2::coord_flip()}.}

\item{lab_other}{character. Name to give to 'opposing' factor of \code{multi_main}.}

\item{lab_title}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_subtitle}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_caption}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_x}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{theme_base}{\code{ggplot2} theme (e.g. as \code{ggplot2::theme_minimal()}.)
A custom \code{tetext} thems is supplied as a default.}

\item{facet_scales}{character. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_ncol}{numeric. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_nrow}{numeric. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_strip_position}{character. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}
}
\value{
gg.
}
\description{
Visualize n-gram log ratios across pairs of \code{multi} values.
}
\details{
\code{compute_logratios_multi_by2_at()} should NOT be called beforehand.
}
\seealso{
\url{https://www.tidytextmining.com/twitter.html#comparing-word-usage}.
\url{http://varianceexplained.org/r/trump-tweets/}.
}
