% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sents.R
\name{visualize_sentratios_facet_by2_at}
\alias{visualize_sentratios_facet_by2_at}
\alias{visualize_sentratios_facet_by2}
\title{Visualize token sentiment log ratios in pairs}
\usage{
visualize_sentratios_facet_by2_at(data = NULL, token = NULL, facet = NULL,
  xy_nms, xy_grid, cnt_min = 10, lexicon = "bing", ...,
  filter_sent = TRUE, sent_main = NULL, filter_facet = TRUE,
  facet_main = NULL, x_include = NULL, y_include = NULL,
  x_exclude = NULL, y_exclude = NULL, color = NULL,
  color_value = c("grey50", "grey80"), num_top = 3, flip_axes = FALSE,
  lab_other = "other",
  title = "Most Significant Words Contributing to Sentiment Differences",
  subtitle = NULL, caption = NULL, lab_x = NULL,
  lab_y = "Log Odds Ratio", theme_base = theme_tetext_facet(),
  scales = "free", ncol = 3, nrow = NULL, strip.position = "right")

visualize_sentratios_facet_by2(data = NULL, token = NULL, facet = NULL,
  xy_nms, xy_grid, cnt_min = 10, lexicon = "bing", ...,
  filter_sent = TRUE, sent_main = NULL, filter_facet = TRUE,
  facet_main = NULL, x_include = NULL, y_include = NULL,
  x_exclude = NULL, y_exclude = NULL, color = NULL,
  color_value = c("grey50", "grey80"), num_top = 3, flip_axes = FALSE,
  lab_other = "other",
  title = "Most Significant Words Contributing to Sentiment Differences",
  subtitle = NULL, caption = NULL, lab_x = NULL,
  lab_y = "Log Odds Ratio", theme_base = theme_tetext_facet(),
  scales = "free", ncol = 3, nrow = NULL, strip.position = "right")
}
\arguments{
\item{data}{data.frame. Data to apply a function over.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use for count. Default is provided}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} corresponding to group
by which count is made.}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{cnt_min}{numeric. If not null, serves as a filter for which to compute log ratios.}

\item{lexicon}{character. Specifies lexicon to use. Must be a lexicon available in the \code{tidytext::package}.}

\item{...}{dots. Additional parameters to pass to \code{compute_sentratios_facet_by2_at()}.}

\item{filter_sent}{logical. Whether or not to filter the sentiment values.}

\item{sent_main}{bare for NSE; character for SE. Name of single sentiment value to use as basis.
Not used if \code{filter_sent = FALSE}.}

\item{filter_facet}{logical. Whether or not to filter the \code{facet} values.}

\item{facet_main}{bare for NSE; character for SE. Name of single \code{facet} value to use as basis.
Not used if \code{filter_facet = FALSE}.}

\item{x_include}{character (vector).
\code{facet} values to include and exclude. Not used if \code{filter_facet = FALSE}.}

\item{y_include}{character (vector).
\code{facet} values to include and exclude. Not used if \code{filter_facet = FALSE}.}

\item{x_exclude}{character (vector).
\code{facet} values to include and exclude. Not used if \code{filter_facet = FALSE}.}

\item{y_exclude}{character (vector).
\code{facet} values to include and exclude. Not used if \code{filter_facet = FALSE}.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{color_value}{character (vector). Should be a vector of length two (for dual colors)
if \code{filter_facet = TRUE} and \code{!is.null(facet_main) & length(facet_main) == 1}.}

\item{num_top}{numeric. Number of tokens to show for each \code{facet} pair.}

\item{flip_axes}{logical. Whether or not to call \code{ggplot2::coord_flip()}.}

\item{lab_other}{bare for NSE; character for SE. Name to give to 'opposing' factor of \code{facet_main}.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}
}
\value{
gg.
}
\description{
Visualize token sentiment log ratios across pairs of \code{facet} values.
}
\details{
\code{compute_sentratios_facet_by2_at()} should NOT be called beforehand.
}
\seealso{
\url{https://www.tidytextmining.com/twitter.html#comparing-token-usage}.
\url{http://varianceexplained.org/r/trump-tweets/}.
\url{https://www.tidytextmining.com/sentiment.html#most-positive-negative}.
}
