% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{visualize_tfidf_at}
\alias{visualize_tfidf_at}
\alias{visualize_tfidf}
\title{Visualize TF-IDF}
\usage{
visualize_tfidf_at(data = NULL, token = NULL, doc = "document",
  facet = doc, num_top = 10, color = facet, color_value = "grey50",
  lab_x = NULL, lab_y = NULL, title = "Highest TF-IDF Words",
  subtitle = NULL, caption = NULL, theme_base = theme_tetext_facet())

visualize_tfidf(data = NULL, token = NULL, doc = "document",
  facet = doc, num_top = 10, color = facet, color_value = "grey50",
  lab_x = NULL, lab_y = NULL, title = "Highest TF-IDF Words",
  subtitle = NULL, caption = NULL, theme_base = theme_tetext_facet())
}
\arguments{
\item{data}{data.frame.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use as unit for counting.
Probably something like 'word' or 'bigram'.}

\item{doc}{bare for NSE; character for SE. Name of column in \code{data} to use
as \code{document} in \code{tidytext::bind_tf_idf()}.}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} to use for facetting.}

\item{num_top}{numeric. Number of words to show. If between 0 and 1, then assumed to be a percentage.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}
}
\value{
gg.
}
\description{
Visualize term-frequency, inverse-document-frequency.
}
\details{
Calls \code{compute_tfidf_at()} internally.
Cannot pass dots to \code{compute_tfidf_at()} internally because parameters
are used in subsequent processing.
}
\seealso{
\url{https://www.tidytextmining.com/tfidf.html}.
\url{https://juliasilge.com/blog/sherlock-holmes-stm/}
}
