% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{visualize_tfidf_at}
\alias{visualize_tfidf_at}
\alias{visualize_tfidf}
\title{Visualize TF-IDF}
\usage{
visualize_tfidf_at(data = NULL, word = "word", doc = "document",
  multi = doc, num_top = 10, color = multi, color_value = "grey50",
  lab_x = NULL, lab_y = NULL, lab_title = "Highest TF-IDF Words",
  lab_subtitle = NULL, lab_caption = NULL,
  theme_base = theme_tetext_facet())

visualize_tfidf(data = NULL, word = "word", doc = "document",
  multi = doc, num_top = 10, color = multi, color_value = "grey50",
  lab_x = NULL, lab_y = NULL, lab_title = "Highest TF-IDF Words",
  lab_subtitle = NULL, lab_caption = NULL,
  theme_base = theme_tetext_facet())
}
\arguments{
\item{data}{data.frame.}

\item{word}{character. Name of column in \code{data} to use for words.
Probably something like 'word' or 'bigram'.}

\item{doc}{character. Name of column in \code{data} to use
as \code{document} in \code{tidytext::bind_tf_idf()}.}

\item{multi}{character. Name of column in \code{data} to use for facetting.}

\item{num_top}{numeric. Number of words to show. If between 0 and 1, then assumed to be a percentage.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function.}

\item{lab_x}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_title}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_subtitle}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{lab_caption}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{theme_base}{\code{ggplot2} theme (e.g. as \code{ggplot2::theme_minimal()}.)
A custom \code{tetext} thems is supplied as a default.}
}
\value{
gg.
}
\description{
Visualize term-frequency, inverse-document-frequency.
}
\details{
Calls \code{compute_tfidf_at()} internally.
Cannot pass dots to \code{compute_tfidf_at()} internally because parameters
are used in subsequent processing.
}
\seealso{
\url{https://www.tidytextmining.com/tfidf.html}.
\url{https://juliasilge.com/blog/sherlock-holmes-stm/}
}
