% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{visualize_tfidf_at}
\alias{visualize_tfidf_at}
\alias{visualize_tfidf}
\title{Visualize TF-IDF}
\usage{
visualize_tfidf_at(..., token = NULL, doc = NULL, facet = doc,
  color = facet, num_top = 10,
  scale_manual_params = default_scale_manual(), labs_base = default_labs(),
  labs_params = list(title = "Highest TF-IDF Tokens"),
  theme_base = default_theme(axis.text.y = ggplot2::element_text(angle = 45,
  hjust = 1), panel.grid.major.y = ggplot2::element_blank()),
  theme_params = list(), facet_base = default_facet(facet),
  facet_params = list())

visualize_tfidf(..., token, doc, facet, color)
}
\arguments{
\item{...}{dots. Parameters to pass directly to \code{visualize_time()}.}

\item{token}{bare for NSE; character for SE. Name of column in \code{data} to use as unit for counting.
Probably something like 'word' or 'bigram'.}

\item{doc}{bare for NSE; character for SE. Name of column in \code{data} to use
as \code{document} in \code{tidytext::bind_tf_idf()}.}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} to use for facetting.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{num_top}{numeric. Number of words to show. If between 0 and 1, then assumed to be a percentage.}

\item{scale_manual_params}{list. Additional parameters to pass to \code{ggplot2::scale_fill_manual()}
or \code{ggplot2::scale_color_manual()} (e.g. \code{breaks}, \code{labels}, or \code{name}).
The default \code{values} specified by \code{scale_manual_base = default_scale_manual()}
can be overwritten a \code{values} argument in this list (or directly with a
\code{values} argument in the \code{scale_manual_base} list.}

\item{labs_base}{\code{ggplot2::labs()} function. Defaults to a pre-determined set of values.
It is recommended NOT to modify this argument; isntead, the \code{_params} argument should
be used for customization.}

\item{labs_params}{list. Additional parameters to pass to \code{ggplot2::labs()} to use in addition
to and/or override the parameters set in the \code{_base} parameter.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}

\item{theme_params}{list. Additional parameters to pass to \code{ggplot2::labs()}.
Should be used in the same manner as \code{labs_params} for customization.}

\item{facet_base}{\code{ggplot2::facet_wrap()} function. An internal function
sets defaults for \code{facets}, \code{scales}, \code{ncol}, \code{nrow}, and \code{strip.position}
As with the \code{labs_base} and \code{theme_base} arguments, it is NOT recommended to set
this argument explicitly; instead, the accompanying \code{_params} argument should be used.}

\item{facet_params}{list. Additional parameters to pass to \code{ggplot2::facet_wrap()}. Parameters
passed to this argument override thos in the \code{_base} argument.}
}
\value{
gg.
}
\description{
Visualize term-frequency, inverse-document-frequency.
}
\details{
Calls \code{compute_tfidf_at()} internally.
Cannot pass dots to \code{compute_tfidf_at()} internally because parameters
are used in subsequent processing.
}
\seealso{
\url{https://www.tidytextmining.com/tfidf.html}.
\url{https://juliasilge.com/blog/sherlock-holmes-stm/}
}
