% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{visualize_time_at}
\alias{visualize_time_at}
\alias{visualize_time}
\alias{visualize_time_facet_at}
\alias{visualize_time_facet}
\title{Visualize by time period.}
\usage{
visualize_time_at(data = NULL, timebin = NULL, color = NULL,
  geom = c("bar", "hist"), add_alpha = FALSE, alpha_range = c(0.25, 1),
  scale_manual_params = default_scale_manual(), labs_base = default_labs(),
  labs_params = list(title = "Count Over Time"),
  theme_base = default_theme_dx(), theme_params = list())

visualize_time(..., timebin, color)

visualize_time_facet_at(..., theme_base = default_theme_facet_dx(),
  facet = NULL, facet_base = default_facet(facet), facet_params = list())

visualize_time_facet(..., timebin, color, facet)
}
\arguments{
\item{data}{data.frame.}

\item{timebin}{bare for NSE; character for SE. Name of column in \code{data} to use for time axis.
Probably something like 'yyyy', 'mm', etc.}

\item{color}{bare for NSE; character for SE. Name of column in \code{data} to use for color basis.
Even if \code{fill} is the actual \code{ggplot2} aesthetic, the internal code will 'coerce' this
column to \code{fill}.
In order to simplify internal code, MUST be specified.}

\item{geom}{character. 'bar' or 'hist'. 'bar' is probably best for everything except \code{Date} objects.}

\item{add_alpha}{logical. Whether or not to use \code{ggplot2::scale_alpha()} based on count.}

\item{alpha_range}{numeric (vector). Direct parameter passed to \code{range} parameter
of \code{ggplot2::scale_alpha()}.}

\item{scale_manual_params}{list. Parameters to pass to \code{ggplot2::scale_fill_manual()} or
\code{ggplot2::scale_color_manual()}, depending on the aesthetic mapping. A default \code{values}
argument is specified, so if overwriting explicitly, then \code{values} should be included in the list.}

\item{labs_base}{\code{ggplot2::labs()} function. Defaults to a pre-determined set of values.
It is recommended NOT to modify this argument; isntead, the \code{_params} argument should
be used for customization.}

\item{labs_params}{list. Additional parameters to pass to \code{ggplot2::labs()} to use in addition
to and/or override the parameters set in the \code{_base} parameter.}

\item{theme_base}{\code{ggplot2::theme()} function. (e.g. as \code{ggplot2::theme_minimal()}.)
A custom theme is supplied as a default. As with the the \code{labs_base} argument, it is NOT
recommended to change this directly.}

\item{theme_params}{list. Additional parameters to pass to \code{ggplot2::labs()}.
Should be used in the same manner as \code{labs_params} for customization.}

\item{...}{dots. Parameters to pass directly to \code{visualize_time()}.}

\item{facet}{bare for NSE; character for SE. Name of column in \code{data} to use for facetting.}

\item{facet_base}{\code{ggplot2::facet_wrap()} function. An internal function
sets defaults for \code{facets}, \code{scales}, \code{ncol}, \code{nrow}, and \code{strip.position}
As with the \code{labs_base} and \code{theme_base} arguments, it is NOT recommended to set
this argument explicitly; instead, the accompanying \code{_params} argument should be used.}

\item{facet_params}{list. Additional parameters to pass to \code{ggplot2::facet_wrap()}. Parameters
passed to this argument override thos in the \code{_base} argument.}
}
\value{
gg
}
\description{
Visualize time-based data over time.
}
\details{
\code{facet} version is for facetting. Lists are used for
customization arguments (instead of dots) because there
are multiple arguments that can be customized.
}
\seealso{
\url{https://juliasilge.com/blog/ten-thousand-tweets/}.
}
