% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{visualize_time_at}
\alias{visualize_time_at}
\alias{visualize_time}
\alias{visualize_time_multi_at}
\alias{visualize_time_multi}
\title{Visualize by time period.}
\usage{
visualize_time_at(data = NULL, timebin = NULL, geom = c("bar", "hist"),
  add_alpha = FALSE, alpha_range = c(0.25, 1), color = NULL,
  color_value = "grey50", lab_x = NULL, lab_y = NULL,
  lab_title = "Count Over Time", lab_subtitle = NULL, lab_caption = NULL,
  theme_base = theme_tetext())

visualize_time(data = NULL, timebin = NULL, geom = c("bar", "hist"),
  add_alpha = FALSE, alpha_range = c(0.25, 1), color = NULL,
  color_value = "grey50", lab_x = NULL, lab_y = NULL,
  lab_title = "Count Over Time", lab_subtitle = NULL, lab_caption = NULL,
  theme_base = theme_tetext())

visualize_time_multi_at(data = NULL, ..., theme_base = theme_tetext_facet(),
  multi = NULL, facet_scales = "free", facet_ncol = 3,
  facet_nrow = NULL, facet_strip_position = "top")

visualize_time_multi(data = NULL, ..., theme_base = theme_tetext_facet(),
  multi = NULL, facet_scales = "free", facet_ncol = 3,
  facet_nrow = NULL, facet_strip_position = "top")
}
\arguments{
\item{data}{data.frame.}

\item{timebin}{character. Name of column in \code{data} to use for time axis.
Probably something like 'yyyy', 'mm', etc.}

\item{geom}{character. 'bar' or 'hist'. 'bar' is probably best for everything except Date objects.}

\item{add_alpha}{logical. Whether or not to use \code{ggplot2::scale_alpha()} based on count.}

\item{alpha_range}{numeric (vector). Direct parameter passed to \code{range} parameter
of \code{ggplot2::scale_alpha()}.}

\item{color}{character. Name of column in \code{data} to use for color basis.
Set to \code{NULL} by default although not actually required in order to simplify internal code.}

\item{color_value}{character (vector). Hex value(s) of color. May be a vector.
Should not be a function.}

\item{lab_title, lab_subtitle, lab_caption, lab_x, lab_y}{character. It is recommended NOT
to use these parameters directly. (Instead, it is recommended that they are modified
'after' the function is called.) Noneheless, they are provided as inputs and assigned
reasonable defaults (e.g. 'Count Over Time')
where appropriate. (Otherwise, they are given the value of NULL.)}

\item{theme_base}{\code{ggplot2} theme (e.g. as \code{ggplot2::theme_minimal()}.)
A custom \code{tetext} thems is supplied as a default.}

\item{...}{dots. Parameters to pass directly to \code{visualize_time_at()}.}

\item{multi}{character. Name of column in \code{data} to use for facetting.}

\item{facet_scales, facet_strip_position}{character. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}

\item{facet_ncol, facet_nrow}{numeric. Direct parameters to analogous \code{ggplot2::facet_wrap()} parameters.}
}
\value{
gg
}
\description{
Visualize time-based data over time.
}
\details{
'multi' version is for facetting.
}
\seealso{
\url{https://juliasilge.com/blog/ten-thousand-tweets/}.
}
