% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{wrapper_func}
\alias{wrapper_func}
\title{'Engine' for '_by2' functions}
\usage{
wrapper_func(data = NULL, func = NULL, xy_grid = NULL, xy_nms = NULL,
  ...)
}
\arguments{
\item{data}{data.frame. Data to apply a function over.}

\item{func}{function. Wrapped by \code{do.call()}.}

\item{xy_grid}{data.frame. Output from \code{create_xy_grid()}.}

\item{xy_nms}{character (vector). Values to use to create a 'grid' of all
combinations of values.}

\item{...}{dots. Additional parameters to pass to \code{func}.}
}
\value{
data.frame.
}
\description{
Apply a function to \code{data} in pairs (over a \code{facet} column).
}
\details{
Should be replaced by a \code{purrr} function in the future.
Must be careful not to pass \code{NULL} in \code{...}.
}
